Add prebuilt Android OpenSSL package for OpenSSL 1.1.1m
Used OpenSSL source package 1.1.1m was changed little bit. All
ANDROID_NDK_HOME strings were replaced with ANDROID_NDK_ROOT in
Configurations/15-android.conf file

Task-number: QTQAINFRA-4864
Task-number: QTBUG-101863
(cherry picked from commit 686481f48c6b3bb18a74616c7ce6a636dd0460cc)

+++ coin/provisioning/common/windows/android-openssl.ps1
@@ -47,47 +47,56 @@ if (Is64BitWinHost) {
 
 $version = "1.1.1m"
 $zip = Get-DownloadLocation ("openssl-$version.tar.gz")
-$prebuilt_zip = Get-DownloadLocation ("openssl-android-master-$version.zip")
-$sha1 = "39d424c4411e45f1570073d7a71b1830b96007ca"
-$prebuilt_sha1 = "07fad2a44ffa90261a779782bd64fe2304487945"
+$prebuilt_zip = Get-DownloadLocation ("prebuilt-openssl-${version}-fixes-ndk_root-windows.zip")
+$sha1 = "c9638d25b9709eda1ac52591c0993af52d6d1206"
+$prebuilt_sha1 = "2bf5354b2264ed80e85fea3705ba434a38fe563e"
 $destination = "C:\Utils\openssl-android-master"
+$prebuilt_url = "\\ci-files01-hki.intra.qt.io\provisioning\openssl\prebuilt-openssl-${version}-fixes-ndk_root-windows.zip"
 
 # msys unix style paths
 $ndkPath = "/c/Utils/Android/android-ndk-r20"
 $openssl_path = "/c/Utils/openssl-android-master"
 $cc_path = "$ndkPath/toolchains/llvm/prebuilt/windows-x86_64/bin"
-Download https://www.openssl.org/source/openssl-$version.tar.gz \\ci-files01-hki.intra.qt.io\provisioning\openssl\openssl-$version.tar.gz $zip
-Verify-Checksum $zip $sha1
 
-Extract-7Zip $zip C:\Utils\tmp
-Extract-7Zip C:\Utils\tmp\openssl-$version.tar C:\Utils\tmp
-Move-Item C:\Utils\tmp\openssl-${version} $destination
-Remove-Item -Path $zip
+if ((Test-Path $prebuilt_url)) {
+    Download $prebuilt_url $prebuilt_url $prebuilt_zip
+    Verify-Checksum $prebuilt_zip $prebuilt_sha1
+    Extract-7Zip $prebuilt_zip C:\Utils
+    Remove $prebuilt_zip
+} else {
+    # openssl-${version}_fixes-ndk_root.tar.gz package includes fixes from https://github.com/openssl/openssl/pull/17322 and string ANDROID_NDK_HOME is replaced with ANDROID_NDK_ROOT in Configurations/15-android.conf
+    Download \\ci-files01-hki.intra.qt.io\provisioning\openssl\openssl-${version}_fixes-ndk_root.tar.gz \\ci-files01-hki.intra.qt.io\provisioning\openssl\openssl-${version}_fixes-ndk_root.tar.gz $zip
+    Verify-Checksum $zip $sha1
+    Extract-7Zip $zip C:\Utils\tmp
+    Extract-7Zip C:\Utils\tmp\openssl-$version.tar C:\Utils\tmp
+    Move-Item C:\Utils\tmp\openssl-${version} $destination
+    Remove-Item -Path $zip
 
-Write-Host "Configuring OpenSSL $version for Android..."
-Push-Location $destination
-# $ must be escaped in powershell...
+    Write-Host "Configuring OpenSSL $version for Android..."
+    Push-Location $destination
+    # $ must be escaped in powershell...
 
-function CheckExitCode {
+    function CheckExitCode {
 
-    param (
-        $p
-    )
+        param (
+            $p
+        )
 
-    if ($p.ExitCode) {
-        Write-host "Process failed with exit code: $($p.ExitCode)"
-        exit 1
+        if ($p.ExitCode) {
+            Write-host "Process failed with exit code: $($p.ExitCode)"
+            exit 1
+        }
     }
+
+    $configure = Start-Process -NoNewWindow -Wait -PassThru -ErrorAction Stop -FilePath "$msys_bash" -ArgumentList ("-lc", "`"pushd $openssl_path; ANDROID_NDK_HOME=$ndkPath PATH=${cc_path}:`$PATH CC=clang $openssl_path/Configure shared android-arm`"")
+    CheckExitCode $configure
+
+    $make = Start-Process -NoNewWindow -Wait -PassThru -ErrorAction Stop -FilePath "$msys_bash" -ArgumentList ("-lc", "`"pushd $openssl_path; ANDROID_NDK_HOME=$ndkPath PATH=${cc_path}:`$PATH CC=clang make -f $openssl_path/Makefile build_generated`"")
+    CheckExitCode $make
+
+    Pop-Location
+    Remove-item C:\Utils\tmp -Recurse -Confirm:$false
 }
 
-$configure = Start-Process -NoNewWindow -Wait -PassThru -ErrorAction Stop -FilePath "$msys_bash" -ArgumentList ("-lc", "`"pushd $openssl_path; ANDROID_NDK_HOME=$ndkPath PATH=${cc_path}:`$PATH CC=clang $openssl_path/Configure shared android-arm`"")
-CheckExitCode $configure
-
-$make = Start-Process -NoNewWindow -Wait -PassThru -ErrorAction Stop -FilePath "$msys_bash" -ArgumentList ("-lc", "`"pushd $openssl_path; ANDROID_NDK_HOME=$ndkPath PATH=${cc_path}:`$PATH CC=clang make -f $openssl_path/Makefile build_generated`"")
-CheckExitCode $make
-
-Pop-Location
-
 Set-EnvironmentVariable "OPENSSL_ANDROID_HOME" "$destination"
-Remove-item C:\Utils\tmp -Recurse -Confirm:$false
 Write-Output "Android OpenSSL = $version" >> ~/versions.txt
